Adding in ProducerConsumer stuff
Before when we we the ProducerConsumer files separate, we were getting a circular definition that resulted in a compiler error.  This is because ProducerConsumer.h included Pcb.h, and Pcb.h included ProducerConsumer.h.  From what I read online, there's no easy way to fix this except merge the two files if we really need them referencing each other.  So that's what I did here.