genProducerConsumerPair
Adding the method (with a helper method - setPCTraps) to generate a producer consumer pair.  Right now it only produces one, there's no loop or anything.  We can either add a loop to this method so it generates as many as we tell it to, or call this method as many times as we want pairs.