more updates
Finished (maybe?) implementing the changes to the layout of PCB.  Namely in here, I've moved some structs from into PCB.h (mutRecPairSteps, prodConsPairSteps, relationship, MRData, and PCData).  Because if I kept them in the .c file, then I was going to have to write a LOT of getters to access the fields inside these structs, and I didn't feel like it.  There are some more getter though, to get the additional fields in PCB (relationship, MRSteps, PCSteps, MRData, and PRData).