more updates
Added in functionality to check (during cpu loop) if the current instruction is a producer/consumer trap call.  If so, handle accordingly.  There's some new helper functions in there, plus "ProdConsTrapHandler" method and a new interrupt type "PRO_CON_INTERRUPT", plus a new case for that interrupt in the scheduler.